refactor(mcp): extract shared describeError helper for tool error handling
Move the error-describing fallback out of automations.service.ts into
frontend/src/mcp/tools/errors.ts and reuse it in ui_navigate, so a
NavigationFailure with no message surfaces its details instead of a
generic fallback string.
